feat: improve sql code, model

+++ person/migrations/0001_add_tables.sql
@@ -1,12 +1,13 @@
 SELECT 'up SQL query';
 CREATE TABLE IF NOT EXISTS persons
 (
-    id         SERIAL PRIMARY KEY ,
+    id         BIGSERIAL PRIMARY KEY ,
     name       VARCHAR(255) NOT NULL,
     surname    VARCHAR(255),
     patronymic VARCHAR(255),
-    age        INTEGER,
+    age        SMALLINT,
     gender     VARCHAR(255)
+
 );
 
 CREATE TABLE IF NOT EXISTS countries
